Report missing signature declaration with multiple return values
The previous message which reported when a function's return signature
didn't declare an error variable didn't work for functions with multiple
return values.

I thought about extending it and coming up with generated variable names
for the non-error types to be able to support automatic fixing, but this
idea ended up being too complicated. Also, I didn't like the idea of an
automated fixer coming up with generated variable names, because you'd
need to manually rename them anyway. That kind of defeats the purpose of
it being automated.

Instead, change the diagnostic message to only refer to the error
variable and recommend what the declared variable name should be based
on the identifier the error was assigned to in the defer.

Modify our multiple return value test functions according to this new
approach.